# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * РемонтPRO API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Manager {
  id: number;
  name: string;
  created_at: string;
  training_count: number;
  completed_count: number;
  /** @nullable */
  avg_score?: number | null;
}

export interface ManagerInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
}

export type ManagerDetailAvgScoresByCriterion = {[key: string]: number | null};

export interface Training {
  id: number;
  manager_id: number;
  manager_name: string;
  difficulty: string;
  property_type: string;
  status: string;
  started_at: string;
  /** @nullable */
  completed_at?: string | null;
  /** @nullable */
  overall_score?: number | null;
}

export interface ManagerDetail {
  id: number;
  name: string;
  created_at: string;
  training_count: number;
  completed_count: number;
  /** @nullable */
  avg_score?: number | null;
  avg_scores_by_criterion: ManagerDetailAvgScoresByCriterion;
  recent_trainings: Training[];
  typical_problems: string[];
}

export interface TrainingInput {
  manager_id: number;
  difficulty: string;
  property_type: string;
}

/**
 * @nullable
 */
export type TrainingDetailClientProfile = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type TrainingDetailFinalReport = { [key: string]: unknown } | null;

export interface Message {
  id: number;
  training_id: number;
  role: string;
  content: string;
  created_at: string;
}

export type FeedbackFeedbackJson = { [key: string]: unknown };

export interface Feedback {
  id: number;
  training_id: number;
  manager_message_id: number;
  feedback_json: FeedbackFeedbackJson;
  created_at: string;
}

export interface TrainingDetail {
  id: number;
  manager_id: number;
  manager_name: string;
  difficulty: string;
  property_type: string;
  status: string;
  started_at: string;
  /** @nullable */
  completed_at?: string | null;
  /** @nullable */
  overall_score?: number | null;
  /** @nullable */
  client_profile?: TrainingDetailClientProfile;
  /** @nullable */
  final_report?: TrainingDetailFinalReport;
  messages: Message[];
  feedbacks: Feedback[];
}

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}

export type AIResponseCoachFeedback = {
  strengths: string;
  mistakes: string;
  next_step: string;
  recommendation: string;
  example_phrase: string;
};

export interface AIResponse {
  client_reply: string;
  coach_feedback: AIResponseCoachFeedback;
  manager_message_id: number;
  client_message_id: number;
  feedback_id: number;
}

export type FinalReportScores = { [key: string]: unknown };

export interface FinalReport {
  overall_score: number;
  scores: FinalReportScores;
  strengths: string[];
  problems: string[];
  recommendations: string[];
  ideal_strategy: string;
  client_outcome: string;
}

export interface AdminLoginInput {
  login: string;
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
  message: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface DeleteResult {
  success: boolean;
  message: string;
}

export interface StatsOverview {
  total_managers: number;
  total_trainings: number;
  completed_trainings: number;
  /** @nullable */
  avg_overall_score?: number | null;
  /** @nullable */
  avg_objections_score?: number | null;
  /** @nullable */
  avg_closing_score?: number | null;
  recent_trainings: Training[];
}

export type StatsDetailedAvgScoresByCriterion = {[key: string]: number | null};

export interface ManagerRankItem {
  manager_id: number;
  manager_name: string;
  /** @nullable */
  avg_score?: number | null;
  completed_count: number;
}

export interface ScoreTrendItem {
  date: string;
  avg_score: number;
  count: number;
}

export interface StatsDetailed {
  total_trainings: number;
  /** @nullable */
  avg_overall_score?: number | null;
  avg_scores_by_criterion: StatsDetailedAvgScoresByCriterion;
  manager_ranking: ManagerRankItem[];
  score_trend: ScoreTrendItem[];
  common_problems: string[];
}

export type GetTrainingsParams = {
/**
 * @nullable
 */
managerId?: number | null;
/**
 * @nullable
 */
difficulty?: string | null;
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
dateFrom?: string | null;
/**
 * @nullable
 */
dateTo?: string | null;
};

export type GetStatsDetailedParams = {
/**
 * @nullable
 */
managerId?: number | null;
/**
 * @nullable
 */
difficulty?: string | null;
/**
 * @nullable
 */
dateFrom?: string | null;
/**
 * @nullable
 */
dateTo?: string | null;
};