# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: РемонтPRO API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /managers:
    get:
      operationId: getManagers
      tags: [managers]
      summary: List all managers
      responses:
        "200":
          description: List of managers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Manager"
    post:
      operationId: createManager
      tags: [managers]
      summary: Create a manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manager"
  /managers/{id}:
    get:
      operationId: getManager
      tags: [managers]
      summary: Get manager by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Manager
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagerDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteManager
      tags: [managers]
      summary: Delete manager
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "400":
          description: Cannot delete
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Unauthorized
...
  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Check admin session
      responses:
        "200":
          description: Admin session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Unauthorized
...
  /stats/detailed:
    get:
      operationId: getStatsDetailed
      tags: [stats]
      summary: Detailed statistics with filters
      parameters:
        - name: managerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: difficulty
          in: query
          schema:
            type: ["string", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Detailed stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsDetailed"
components:
  schemas:
...
    Manager:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        created_at:
          type: string
        training_count:
          type: integer
        completed_count:
          type: integer
        avg_score:
          type: ["number", "null"]
      required:
        - id
        - name
        - created_at
        - training_count
        - completed_count
    ManagerInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
      required:
        - name
    ManagerDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        created_at:
          type: string
        training_count:
          type: integer
        completed_count:
          type: integer
        avg_score:
          type: ["number", "null"]
        avg_scores_by_criterion:
          type: object
          additionalProperties:
            type: ["number", "null"]
        recent_trainings:
          type: array
          items:
            $ref: "#/components/schemas/Training"
        typical_problems:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - created_at
        - training_count
        - completed_count
        - avg_scores_by_criterion
        - recent_trainings
        - typical_problems
    Training:
      type: object
      properties:
        id:
          type: integer
        manager_id:
          type: integer
        manager_name:
          type: string
        difficulty:
          type: string
        property_type:
          type: string
        status:
          type: string
        started_at:
          type: string
        completed_at:
          type: ["string", "null"]
        overall_score:
          type: ["number", "null"]
      required:
        - id
        - manager_id
        - manager_name
        - difficulty
        - property_type
        - status
        - started_at
...
    AdminLoginInput:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      required:
        - login
        - password
    AdminLoginResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    AdminSession:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    StatsDetailed:
      type: object
      properties:
        total_trainings:
          type: integer
        avg_overall_score:
          type: ["number", "null"]
        avg_scores_by_criterion:
          type: object
          additionalProperties:
            type: ["number", "null"]
        manager_ranking:
          type: array
          items:
            $ref: "#/components/schemas/ManagerRankItem"
        score_trend:
          type: array
          items:
            $ref: "#/components/schemas/ScoreTrendItem"
        common_problems:
          type: array
          items:
            type: string
      required:
        - total_trainings
        - avg_scores_by_criterion
        - manager_ranking
        - score_trend
        - common_problems
    ManagerRankItem:
      type: object
      properties:
        manager_id:
          type: integer
        manager_name:
          type: string
        avg_score:
          type: ["number", "null"]
        completed_count:
          type: integer
      required:
        - manager_id
        - manager_name
        - completed_count
    ScoreTrendItem:
      type: object
      properties:
        date:
          type: string
        avg_score:
          type: number
        count:
          type: integer
      required:
        - date
        - avg_score
        - count