# lib/api-zod/src/generated/types/managerRankItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * РемонтPRO API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ManagerRankItem {
  manager_id: number;
  manager_name: string;
  /** @nullable */
  avg_score?: number | null;
  completed_count: number;
}