# lib/api-zod/src/generated/types/scoreTrendItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * РемонтPRO API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreTrendItem {
  date: string;
  avg_score: number;
  count: number;
}